Moves device query script
