+++ device_query/device_query.py
@@ -1,86 +1,144 @@
-import argparse
-import subprocess
-import xml.etree.ElementTree as ET
-import os
-import pwd
-import re
+import configparser
 import json
-from http.server import BaseHTTPRequestHandler, HTTPServer
+import pwd
+import socket
+import sys
+from time import sleep
+
+import requests
+import xml.etree.ElementTree as ET
 
 
-class DeviceQueryHandler(BaseHTTPRequestHandler):
+def parse_nvidia_xml(xml):
+    gpu_data = []
+    root = ET.fromstring(xml)
 
-    def parse_nvidia_xml(self, xml):
-        gpu_data = []
-        root = ET.fromstring(xml)
+    for gpu in root.iter('gpu'):
+        current_gpu_data = {
+            "name": gpu.find("product_name").text,
+            "uuid": gpu.find("uuid").text,
+        }
 
-        for gpu in root.iter('gpu'):
-            current_gpu_data = {
-                "name": gpu.find("product_name").text,
-                "uuid": gpu.find("uuid").text,
-            }
+        memory_usage = gpu.find("fb_memory_usage")
+        memory = {
+            "total": memory_usage.find("total").text,
+            "used": memory_usage.find("used").text,
+            "free": memory_usage.find("free").text,
+        }
 
-            memory_usage = gpu.find("fb_memory_usage")
-            memory = {
-                "total": memory_usage.find("total").text,
-                "used": memory_usage.find("used").text,
-                "free": memory_usage.find("free").text,
-            }
-
-            process_block = gpu.find("processes")
-            if process_block.text == "N/A":
-                current_gpu_data["in_use"] = "na"
-            else:
-                current_gpu_data["in_use"] = "no"
-                current_gpu_data["processes"] = []
-                for process in process_block.iter("process_info"):
-                    if process.find('type').text.lower() == "c":
-                        current_gpu_data["in_use"] = "yes"
-                        pid = process.find('pid').text
-                        process_info = {
-                            "pid": pid,
-                            "username": self.owner(pid),
-                            "name": process.find("process_name").text,
-                            "used_memory": process.find("used_memory").text,
-                        }
-                        current_gpu_data["processes"].append(process_info)
-            current_gpu_data["memory"] = memory
-            gpu_data.append(current_gpu_data)
-        return gpu_data
-
-    def owner(self, pid):
-        UID = 1
-        for line in open('/proc/{}/status'.format(pid)):
-            if line.startswith('Uid:'):
-                uid = int(line.split()[UID])
-                return pwd.getpwuid(uid).pw_name
-
-    def do_GET(self):
-        try:
-            if not re.match(self.allowed_client, self.client_address[0]):
-                self.send_error(403)
-                return
-
-            raw_gpu_data = subprocess.check_output(["nvidia-smi", "-x", "-q"]).decode('utf-8')
-            gpu_data = self.parse_nvidia_xml(raw_gpu_data)
-            self.send_response(200)
-            self.send_header("Content-type", "application/json")
-            self.end_headers()
-            self.wfile.write(bytes(json.dumps(gpu_data, indent=4), 'utf-8'))
-        except Exception as e:
-            self.send_error(500)
+        process_block = gpu.find("processes")
+        if process_block.text == "N/A":
+            current_gpu_data["in_use"] = "na"
+        else:
+            current_gpu_data["in_use"] = "no"
+            current_gpu_data["processes"] = []
+            for process in process_block.iter("process_info"):
+                if process.find('type').text.lower() == "c":
+                    current_gpu_data["in_use"] = "yes"
+                    pid = process.find('pid').text
+                    process_info = {
+                        "pid": pid,
+                        "username": get_owner_for_pid(pid),
+                        "name": process.find("process_name").text,
+                        "used_memory": process.find("used_memory").text,
+                    }
+                    current_gpu_data["processes"].append(process_info)
+        current_gpu_data["memory"] = memory
+        gpu_data.append(current_gpu_data)
+    return gpu_data
 
 
-if __name__ == "__main__":
-    parser = argparse.ArgumentParser(description='Tool that provides information about GPUs in this machine')
-    parser.add_argument("-ac", "--allowed-client-address", default='.*', required=False, help="Restricts possible clients to given ip address")
+def get_owner_for_pid(pid):
+    UID = 1
+    for line in open('/proc/{}/status'.format(pid)):
+        if line.startswith('Uid:'):
+            uid = int(line.split()[UID])
+            return pwd.getpwuid(uid).pw_name
 
-    args = parser.parse_args()
 
-    RestrictedDeviceQueryHandler = type('RestrictedDeviceQueryHandler', (DeviceQueryHandler,), dict(allowed_client=args.allowed_client_address))
-    server = HTTPServer(("0.0.0.0", 12000), RestrictedDeviceQueryHandler)
-    try:
-        server.serve_forever()
-    except KeyboardInterrupt:
-        pass
-    server.server_close()
+def main():
+    # TODO: check if this still works with the provided service conf
+    # TODO: parse GPU utilization
+    config = configparser.ConfigParser()
+    config.read("config.ini")
+
+    server_url = config["MAIN"]["server_url"]  # TODO: https - in the end
+    update_interval = int(config["MAIN"]["update_interval"])
+    device_name = config["MAIN"]["device_name"]
+
+    auth_token = config["MAIN"]["token"]
+    if auth_token == "":
+        print("Authentication token must be manually set in config.ini file.")
+        sys.exit(1)
+    headers = {"Authorization": f"Token {auth_token}"}
+
+    with open("nvidia-smi-output.xml", "r") as nvidia_data_file:
+        raw_gpu_data = nvidia_data_file.read()
+    # TODO: test proper parser
+    # raw_gpu_data = subprocess.check_output(["nvidia-smi", "-x", "-q"]).decode('utf-8')
+    gpu_data = parse_nvidia_xml(raw_gpu_data)
+
+    #### for testing
+    # gpu_data = [{
+    #     'name': 'Testy GPU',
+    #     'uuid': 'loremuuuuuu',
+    #     'memory': {
+    #         'total': '200 MB',
+    #         'used': '20 MB',
+    #         'free': '180 MB'
+    #     },
+    #     'in_use': 'yes',
+    #     'processes': [{
+    #         'pid': 1,
+    #         'username': 'Mrs. Keksi',
+    #         'name': 'TestProcess',
+    #         'used_memory': '10 MB'
+    #     }]
+    # }]
+    # device_name = "test_device"
+    #
+    # gpu_data = [{
+    #     'name': 'Test GPU',
+    #     'uuid': 'lorem',
+    #     'memory': {
+    #         'total': '100 MB',
+    #         'used': '0 MB',
+    #         'free': '100 MB'
+    #     },
+    #     'in_use': 'no',
+    #     'processes': []
+    # }]
+    # device_name = "test_device2"
+    #
+    # gpu_data = [{
+    #     'name': 'Toast GPU',
+    #     'uuid': 'loremuuuuuu',
+    #     'memory': {
+    #         'total': '200 MB',
+    #         'used': '20 MB',
+    #         'free': '180 MB'
+    #     },
+    #     'in_use': 'yes',
+    #     'processes': [{
+    #         'pid': 1,
+    #         'username': 'Mrs. Keksi',
+    #         'name': 'TestProcess',
+    #         'used_memory': '10 MB'
+    #     }]
+    # }]
+    # device_name = "test_device"
+    #### testing end
+
+    post_data = {
+        "gpu_data": gpu_data,
+        "device_name": device_name
+    }
+    encoded_post_data = bytes(json.dumps(post_data, indent=4), 'utf-8')
+    while True:
+        r = requests.post(server_url, headers=headers, data=encoded_post_data)
+        sleep(update_interval)
+        break  # TODO remove
+
+
+if __name__ == '__main__':
+    main()
